:recycle: refactor(component): update state hoisting and naming
