+++ app/src/main/java/com/anna/healthyeats/ui/components/forms/email/EmailFieldPreviews.kt
@@ -1,37 +1,31 @@
-package com.anna.healthyeats.ui.components.forms.text
+package com.anna.healthyeats.ui.components.forms.email
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
 import com.anna.healthyeats.R
 import com.anna.healthyeats.ui.components.forms.common.HealthyEatsFieldError
 import com.anna.healthyeats.ui.theme.HealthyEatsTheme
 
 /**====================================================================
- * Text field Previews
+ * Email field previews
  * ===================================================================*/
 @Composable
 @Preview(showBackground = true, showSystemUi = true, group = "filled")
-fun HealthyEatsFilledTextFieldPreview(){
-  var input by remember { mutableStateOf("Anna") }
+fun HealthyEatsFilledEmailFieldPreview(){
+  val input = remember { mutableStateOf("") }
 
   HealthyEatsTheme {
     Column {
-      HealthyEatsTextField(
-        input = input,
-        placeholder = stringResource(id = R.string.name_placeholder),
-        onInputChange = {
-          input = it
-        },
+      HealthyEatsEmailField(
+        inputState = input,
+        placeholder = stringResource(id = R.string.email_placeholder),
         modifier = Modifier.padding(dimensionResource(id = R.dimen.screen_medium_padding))
       )
     }
@@ -40,23 +34,22 @@ fun HealthyEatsFilledTextFieldPreview(){
 
 @Composable
 @Preview(showBackground = true, showSystemUi = true, group = "empty")
-fun HealthyEatsEmptyTextFieldPreview(){
-  var input by remember { mutableStateOf("") }
+fun HealthyEatsEmptyEmailFieldPreview(){
+  val input = remember { mutableStateOf("") }
 
   HealthyEatsTheme {
+    val errorMessage = stringResource(id = R.string.email_required_error)
+
     Column {
-      HealthyEatsTextField(
-        input = input,
-        placeholder = stringResource(id = R.string.name_placeholder),
-        onInputChange = {
-          input = it
-        },
+      HealthyEatsEmailField(
+        inputState = input,
+        placeholder = stringResource(id = R.string.email_placeholder),
         modifier = Modifier.padding(dimensionResource(id = R.dimen.screen_medium_padding)),
         isError = true,
-        errorMessage = stringResource(id = R.string.name_required_error)
+        errorMessage = errorMessage
       )
 
-      HealthyEatsFieldError(modifier = Modifier.padding(start = 16.dp), errorMessage = stringResource(id = R.string.name_required_error))
+      HealthyEatsFieldError(modifier = Modifier.padding(start = dimensionResource(id = R.dimen.screen_medium_padding)), errorMessage = errorMessage)
     }
   }
 }
