+++ app/src/main/java/com/anna/healthyeats/ui/components/forms/password/PasswordField.kt
@@ -1,25 +1,34 @@
-package com.anna.healthyeats.ui.components.forms.text
+package com.anna.healthyeats.ui.components.forms.password
 
 import androidx.compose.foundation.text.KeyboardActionScope
 import androidx.compose.foundation.text.KeyboardActions
+import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.MutableState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalSoftwareKeyboardController
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.text.input.VisualTransformation
 import com.anna.healthyeats.ui.components.forms.common.ErrorTrailingIcon
 import com.anna.healthyeats.ui.components.forms.common.KeyOptions
+import com.anna.healthyeats.ui.components.forms.common.PasswordLeadingIcon
+import com.anna.healthyeats.ui.components.forms.common.PasswordTrailingIcon
 import com.anna.healthyeats.ui.components.forms.common.fieldColorScheme
 import com.anna.healthyeats.ui.components.forms.common.getKeyboard
 import com.anna.healthyeats.ui.components.forms.common.healthyEatsField
 
 /**
- * Normal text field
- * @param input Field value
+ * Password field
+ * @param inputState Field state
  * @param placeholder Text to be displayed when no value is present
- * @param onInputChange Function/Method to run when the field value changes
  * @param modifier Modifier instance to add styling
  * @param isError Whether field validation has failed
  * @param errorMessage Error message to be displayed
@@ -27,28 +36,39 @@ import com.anna.healthyeats.ui.components.forms.common.healthyEatsField
  */
 @OptIn(ExperimentalComposeUiApi::class)
 @Composable
-fun HealthyEatsTextField(
-  input: String,
+fun HealthyEatsPasswordField(
+  inputState: MutableState<String>,
   placeholder: String,
-  onInputChange: (String) -> Unit,
   modifier: Modifier,
   isError: Boolean?= false,
   errorMessage: String?= "",
   onDone: (KeyboardActionScope.() -> Unit)?= null,
 ){
   val keyboardController = LocalSoftwareKeyboardController.current
+  var passwordHidden by rememberSaveable { mutableStateOf(true) }
 
   OutlinedTextField(
-    value = input,
-    onValueChange = onInputChange,
+    value = inputState.value,
+    onValueChange = { newValue ->
+      inputState.value = newValue
+    },
     modifier = Modifier.healthyEatsField(isError, errorMessage, modifier),
     singleLine = true,
     textStyle = MaterialTheme.typography.bodyMedium,
     isError = isError as Boolean,
     placeholder = { Text(text = placeholder) },
-    trailingIcon = { if(isError) ErrorTrailingIcon() },
+    leadingIcon = { PasswordLeadingIcon() },
+    trailingIcon = {
+      if (!isError) {
+        IconButton(onClick = { passwordHidden = !passwordHidden }) {
+          PasswordTrailingIcon(passwordHidden = passwordHidden)
+        }
+      } else {
+        ErrorTrailingIcon()
+      }
+    },
     colors = fieldColorScheme(),
-    keyboardOptions = getKeyboard(KeyOptions.TEXT_INPUT),
+    keyboardOptions = getKeyboard(KeyOptions.PASSWORD_INPUT),
     keyboardActions = KeyboardActions(
       onDone = {
         if(onDone is (KeyboardActionScope.() -> Unit)){
@@ -57,6 +77,7 @@ fun HealthyEatsTextField(
           keyboardController?.hide()
         }
       }
-    )
+    ),
+    visualTransformation = if (passwordHidden) PasswordVisualTransformation() else VisualTransformation.None
   )
 }
