+++ app/src/main/java/com/anna/healthyeats/ui/components/forms/search/SearchBar.kt
@@ -4,33 +4,35 @@ import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.SearchBar
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.MutableState
 import androidx.compose.ui.Modifier
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun HealthyEatsSearchBar(
-  query: String,
-  onQueryChange: (String) -> Unit,
-  onSearch: (String) -> Unit,
-  active: Boolean,
-  onActiveChange: (Boolean) -> Unit,
+  queryState: MutableState<String>,
+  activeState: MutableState<Boolean>,
   placeholder: String,
   modifier: Modifier = Modifier,
   content: @Composable (() -> Unit)
 ){
   SearchBar(
-    query = query,
+    query = queryState.value,
     modifier = Modifier.healthyEatsSearch(modifier),
-    onQueryChange = onQueryChange,
-    onSearch = onSearch ,
-    active = active,
-    onActiveChange = onActiveChange,
+    onQueryChange = { newValue ->
+      queryState.value = newValue
+    },
+    onSearch = { activeState.value = false } ,
+    active = activeState.value,
+    onActiveChange = { newActiveValue ->
+      activeState.value = newActiveValue
+    },
     colors = searchColorScheme(),
     placeholder = { Text(text = placeholder) },
     leadingIcon = { SearchLeadingIcon() },
     trailingIcon = { SearchTrailingIcon() }
     ) {
-    content
+    content()
   }
 }
