+++ app/src/main/java/com/anna/healthyeats/ui/components/forms/text/TextField.kt
@@ -1,34 +1,25 @@
-package com.anna.healthyeats.ui.components.forms.password
+package com.anna.healthyeats.ui.components.forms.text
 
 import androidx.compose.foundation.text.KeyboardActionScope
 import androidx.compose.foundation.text.KeyboardActions
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
+import androidx.compose.runtime.MutableState
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalSoftwareKeyboardController
-import androidx.compose.ui.text.input.PasswordVisualTransformation
-import androidx.compose.ui.text.input.VisualTransformation
 import com.anna.healthyeats.ui.components.forms.common.ErrorTrailingIcon
 import com.anna.healthyeats.ui.components.forms.common.KeyOptions
-import com.anna.healthyeats.ui.components.forms.common.PasswordLeadingIcon
-import com.anna.healthyeats.ui.components.forms.common.PasswordTrailingIcon
 import com.anna.healthyeats.ui.components.forms.common.fieldColorScheme
 import com.anna.healthyeats.ui.components.forms.common.getKeyboard
 import com.anna.healthyeats.ui.components.forms.common.healthyEatsField
 
 /**
- * Password field
- * @param input Field value
+ * Normal text field
+ * @param inputState Field state
  * @param placeholder Text to be displayed when no value is present
- * @param onInputChange Function/Method to run when the field value changes
  * @param modifier Modifier instance to add styling
  * @param isError Whether field validation has failed
  * @param errorMessage Error message to be displayed
@@ -36,38 +27,29 @@ import com.anna.healthyeats.ui.components.forms.common.healthyEatsField
  */
 @OptIn(ExperimentalComposeUiApi::class)
 @Composable
-fun HealthyEatsPasswordField(
-  input: String,
+fun HealthyEatsTextField(
+  inputState: MutableState<String>,
   placeholder: String,
-  onInputChange: (String) -> Unit,
   modifier: Modifier,
   isError: Boolean?= false,
   errorMessage: String?= "",
   onDone: (KeyboardActionScope.() -> Unit)?= null,
 ){
   val keyboardController = LocalSoftwareKeyboardController.current
-  var passwordHidden by rememberSaveable { mutableStateOf(true) }
 
   OutlinedTextField(
-    value = input,
-    onValueChange = onInputChange,
+    value = inputState.value,
+    onValueChange = { newValue ->
+      inputState.value = newValue
+    },
     modifier = Modifier.healthyEatsField(isError, errorMessage, modifier),
     singleLine = true,
     textStyle = MaterialTheme.typography.bodyMedium,
     isError = isError as Boolean,
     placeholder = { Text(text = placeholder) },
-    leadingIcon = { PasswordLeadingIcon() },
-    trailingIcon = {
-      if (!isError) {
-        IconButton(onClick = { passwordHidden = !passwordHidden }) {
-          PasswordTrailingIcon(passwordHidden = passwordHidden)
-        }
-      } else {
-        ErrorTrailingIcon()
-      }
-    },
+    trailingIcon = { if(isError) ErrorTrailingIcon() },
     colors = fieldColorScheme(),
-    keyboardOptions = getKeyboard(KeyOptions.PASSWORD_INPUT),
+    keyboardOptions = getKeyboard(KeyOptions.TEXT_INPUT),
     keyboardActions = KeyboardActions(
       onDone = {
         if(onDone is (KeyboardActionScope.() -> Unit)){
@@ -76,7 +58,6 @@ fun HealthyEatsPasswordField(
           keyboardController?.hide()
         }
       }
-    ),
-    visualTransformation = if (passwordHidden) PasswordVisualTransformation() else VisualTransformation.None
+    )
   )
 }
