+++ app/src/main/java/com/anna/healthyeats/ui/components/forms/text/TextFieldPreviews.kt
@@ -1,36 +1,32 @@
-package com.anna.healthyeats.ui.components.forms.password
+package com.anna.healthyeats.ui.components.forms.text
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
 import com.anna.healthyeats.R
 import com.anna.healthyeats.ui.components.forms.common.HealthyEatsFieldError
 import com.anna.healthyeats.ui.theme.HealthyEatsTheme
 
 /**====================================================================
- * Password field previews
+ * Text field Previews
  * ===================================================================*/
 @Composable
 @Preview(showBackground = true, showSystemUi = true, group = "filled")
-fun HealthyEatsFilledPasswordFieldPreview(){
-  var input by remember { mutableStateOf("12345678") }
+fun HealthyEatsFilledTextFieldPreview(){
+  val input = remember { mutableStateOf("Anna") }
 
   HealthyEatsTheme {
     Column {
-      HealthyEatsPasswordField(
-        input = input,
-        placeholder = stringResource(id = R.string.password_placeholder),
-        onInputChange = {
-          input = it
-        },
+      HealthyEatsTextField(
+        inputState = input,
+        placeholder = stringResource(id = R.string.name_placeholder),
         modifier = Modifier.padding(dimensionResource(id = R.dimen.screen_medium_padding))
       )
     }
@@ -39,23 +35,20 @@ fun HealthyEatsFilledPasswordFieldPreview(){
 
 @Composable
 @Preview(showBackground = true, showSystemUi = true, group = "empty")
-fun HealthyEatsEmptyPasswordFieldPreview(){
-  var input by remember { mutableStateOf("") }
+fun HealthyEatsEmptyTextFieldPreview(){
+  val input = remember { mutableStateOf("") }
 
   HealthyEatsTheme {
     Column {
-      HealthyEatsPasswordField(
-        input = input,
-        placeholder = stringResource(id = R.string.password_placeholder),
-        onInputChange = {
-          input = it
-        },
+      HealthyEatsTextField(
+        inputState = input,
+        placeholder = stringResource(id = R.string.name_placeholder),
         modifier = Modifier.padding(dimensionResource(id = R.dimen.screen_medium_padding)),
         isError = true,
-        errorMessage = stringResource(id = R.string.password_placeholder)
+        errorMessage = stringResource(id = R.string.name_required_error)
       )
 
-      HealthyEatsFieldError(modifier = Modifier.padding(start = dimensionResource(id = R.dimen.screen_medium_padding)), errorMessage = stringResource(id = R.string.password_required_error))
+      HealthyEatsFieldError(modifier = Modifier.padding(start = 16.dp), errorMessage = stringResource(id = R.string.name_required_error))
     }
   }
 }
