+++ app/src/main/java/com/anna/healthyeats/ui/components/layout/bottom_sheet/BottomSheet.kt
@@ -1,13 +1,10 @@
 package com.anna.healthyeats.ui.components.layout.bottom_sheet
 
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.ColumnScope
-import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.BottomSheetScaffold
 import androidx.compose.material3.BottomSheetScaffoldState
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.Dp
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -15,7 +12,6 @@ import androidx.compose.ui.unit.Dp
 fun HealthyEatsBottomSheet(
   scaffoldState: BottomSheetScaffoldState,
   sheetContent: @Composable (ColumnScope.() -> Unit),
-  modifier: Modifier = Modifier,
   sheetPeekHeight: Dp ?= null,
   content: @Composable (() -> Unit)
 ){
@@ -25,9 +21,7 @@ fun HealthyEatsBottomSheet(
     sheetContent = sheetContent,
     sheetContainerColor = bottomSheetDefaults().sheetContainerColor,
     containerColor = bottomSheetDefaults().containerColor,
-    sheetDragHandle = { HealthyEatsBottomSheetDragHandle() }) { innerPadding ->
-    Box(modifier.padding(innerPadding)) {
-      content
-    }
+    sheetDragHandle = { HealthyEatsBottomSheetDragHandle() }) {
+       content()
    }
 }
